+++ src/EntityMaxLengthTrim/Extensions/ObjectExtensions.cs
@@ -37,5 +37,16 @@ namespace EntityMaxLengthTrim.Extensions
         {
             return source == null;
         }
+
+        /// <summary>
+        ///     Check if the source object is not null
+        /// </summary>
+        /// <param name="source">Object to be checked</param>
+        /// <returns>Return bool value (validation result).</returns>
+        [CodeSource("https://github.com/I-RzR-I/DomainCommonExtensions", "RzR", 1.0)]
+        internal static bool IsNotNull(this object source)
+        {
+            return !source.IsNull();
+        }
     }
 }